コンポーネントの型定義を更新し、日付表示のロジックを改善しました。CarouselSectionでのオポチュニティの型を拡張し、複数のコンポーネントで日付フォーマット関数を統一しました。また、QuestListのタイトルを修正し、ポイント表示の形式を最適化しました。
