+++ src/app/activities/components/CarouselSection/CarouselSection.tsx
@@ -2,12 +2,13 @@
 
 import React from "react";
 import OpportunityCardVertical from "@/app/activities/components/Card/CardVertical";
-import { ActivityCard } from "@/app/activities/data/type";
+import { ActivityCard, QuestCard } from "@/app/activities/data/type";
 import CarouselSectionSkeleton from "@/app/activities/components/CarouselSection/CarouselSectionSkeleton";
 
+type CardType = ActivityCard | QuestCard;
 interface ActivitiesCarouselSectionProps {
   title: string;
-  opportunities: ActivityCard[];
+  opportunities: CardType[];
   isInitialLoading?: boolean;
   isSearchResult?: boolean;
 }
