+++ src/app/components/CardHorizontal.tsx
@@ -6,7 +6,7 @@ import Image from "next/image";
 import React, { useCallback, useState } from "react";
 import { PLACEHOLDER_IMAGE } from "@/utils";
 import { GqlOpportunityCategory } from "@/types/graphql";
-import { formatDateTimeFromISO } from "@/utils/date";
+import { displayDuration } from "@/utils/date";
 import { Button } from "@/components/ui/button";
 import SelectionSheet from "../reservation/select-date/components/SelectionSheet";
 
@@ -59,7 +59,7 @@ export const CardHorizontal = ({ opportunity, withShadow = true, category, start
           <div>
             <h2 className="text-label-sm font-bold pt-10">日時</h2>
             <p className="text-body-sm text-foreground pt-1">
-              {formatDateTimeFromISO(startDateTime?.toISOString() ?? "", endDateTime?.toISOString() ?? "")}
+              {displayDuration(startDateTime?.toISOString() ?? "", endDateTime?.toISOString() ?? "")}
             </p>
           </div>
           <div className="border-b border-gray-200 my-4"></div>
